Create function that removes elements from array
Create a new array, and use for loop that loops through all instances
of arguments that are present. If statement that if there are
arguments that aren't in the array, then move to the next iteration.
If there is, then filter out all instances of that argument in the
array and return the new array.