Create function that adds all numbers between
Function will see if the numbers are both integers and positive,
then it will sum all the numbers in between and including the numbers
and returns that number. Use Math.max and Math.min to find the smallest
of the two paramters. Then use for loop to iterate through the numbers
in between. Use Number.isInteger() and greater than sign to check if
number is an integer and positive.