Create function that tells if leap leapYears
if statement that sees if number is divisible by 4, followed
by another if statement that checks to see if that number is
also divisible by 100, and also needs to be divisible by 400. If both
are not true, then returns true. if divisble by 100 is true, and it
cant be divisble by 400, then it's not a leap year.