Create function that finds oldest person in array
Use sort method. Define a constant of the date to get the year from.
Define variables that return whether they have a year of death, if they
dont then assign their year of death to this year. Then sort through
their age and have the oldest person at the beginning of array. Then
return the first/beginning object of array.